Edvisr-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.11.3
python-dotenv==1.2.1
slowapi
alembic
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
python-multipart==0.0.20
email-validator==2.2.0
google-genai>=1.0.0
PyJWT==2.10.1
